Moved test data and schema
