+++ projects/_03_independent_cqc/_06_estimate_filled_posts/tests/utils/test_estimate_ind_cqc_filled_posts_models_utils.py
@@ -14,8 +14,12 @@ from utils.column_values.categorical_column_values import CareHome
 from projects._03_independent_cqc._06_estimate_filled_posts.utils.models.utils import (
     utils as job,
 )
-from tests.test_file_data import EstimateFilledPostsModelsUtils as Data
-from tests.test_file_schemas import EstimateFilledPostsModelsUtils as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    EstimateFilledPostsModelsUtils as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    EstimateFilledPostsModelsUtils as Schemas,
+)
 
 
 PATCH_PATH: str = (
