+++ projects/_03_independent_cqc/_06_estimate_filled_posts/tests/utils/test_extrapolation.py
@@ -9,8 +9,12 @@ from utils.column_names.ind_cqc_pipeline_columns import (
 )
 
 import projects._03_independent_cqc._06_estimate_filled_posts.utils.models.extrapolation as job
-from tests.test_file_data import ModelExtrapolation as Data
-from tests.test_file_schemas import ModelExtrapolation as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    ModelExtrapolation as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    ModelExtrapolation as Schemas,
+)
 
 
 PATCH_PATH = (
