+++ projects/_03_independent_cqc/_06_estimate_filled_posts/tests/utils/test_imputation_with_extrapolation_and_interpolation.py
@@ -6,10 +6,10 @@ from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCqc
 
 import projects._03_independent_cqc._06_estimate_filled_posts.utils.models.imputation_with_extrapolation_and_interpolation as job
-from tests.test_file_data import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
     ModelImputationWithExtrapolationAndInterpolationData as Data,
 )
-from tests.test_file_schemas import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
     ModelImputationWithExtrapolationAndInterpolationSchemas as Schemas,
 )
