+++ projects/_03_independent_cqc/_06_estimate_filled_posts/tests/utils/test_ml_model_metrics.py
@@ -8,8 +8,12 @@ from utils.column_names.ind_cqc_pipeline_columns import (
 )
 
 import projects._03_independent_cqc._06_estimate_filled_posts.utils.ml_model_metrics as job
-from tests.test_file_data import MLModelMetrics as Data
-from tests.test_file_schemas import MLModelMetrics as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    MLModelMetrics as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    MLModelMetrics as Schemas,
+)
 
 
 PATCH_PATH = (
