+++ projects/_03_independent_cqc/_06_estimate_filled_posts/tests/utils/test_model_non_res_with_dormancy.py
@@ -7,8 +7,12 @@ from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCqc
 
 import projects._03_independent_cqc._06_estimate_filled_posts.utils.models.non_res_with_dormancy as job
-from tests.test_file_data import ModelNonResWithDormancy as Data
-from tests.test_file_schemas import ModelNonResWithDormancy as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    ModelNonResWithDormancy as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    ModelNonResWithDormancy as Schemas,
+)
 
 
 PATCH_PATH: str = "projects._03_independent_cqc._06_estimate_filled_posts.utils.models.non_res_with_dormancy"
