+++ projects/_03_independent_cqc/_06_estimate_filled_posts/tests/utils/test_model_primary_service_rate_of_change.py
@@ -6,8 +6,12 @@ from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCqc
 
 import projects._03_independent_cqc._06_estimate_filled_posts.utils.models.primary_service_rate_of_change as job
-from tests.test_file_data import ModelPrimaryServiceRateOfChange as Data
-from tests.test_file_schemas import ModelPrimaryServiceRateOfChange as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    ModelPrimaryServiceRateOfChange as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    ModelPrimaryServiceRateOfChange as Schemas,
+)
 
 
 PATCH_PATH: str = "projects._03_independent_cqc._06_estimate_filled_posts.utils.models.primary_service_rate_of_change"
