+++ tests/test_file_data.py
@@ -1557,1110 +1557,6 @@ class EstimateIndCQCFilledPostsData:
 
 
 @dataclass
-class ModelPrimaryServiceRateOfChange:
-    # fmt: off
-    primary_service_rate_of_change_rows = [
-        ("1-001", CareHome.care_home, 1704067200, PrimaryServiceType.care_home_only, 10, 3.0),
-        ("1-001", CareHome.care_home, 1704153600, PrimaryServiceType.care_home_only, 10, 2.8),
-        ("1-001", CareHome.care_home, 1704240000, PrimaryServiceType.care_home_only, 10, 3.4),
-        ("1-001", CareHome.care_home, 1704326400, PrimaryServiceType.care_home_only, 10, 3.2),
-        ("1-002", CareHome.care_home, 1704067200, PrimaryServiceType.care_home_only, 10, 2.0),
-        ("1-002", CareHome.care_home, 1704153600, PrimaryServiceType.care_home_only, 10, None),
-        ("1-002", CareHome.care_home, 1704240000, PrimaryServiceType.care_home_only, 10, None),
-        ("1-002", CareHome.care_home, 1704326400, PrimaryServiceType.care_home_only, 10, 3.2),
-        ("1-003", CareHome.not_care_home, 1704067200, PrimaryServiceType.non_residential, None, 40.0),
-        ("1-003", CareHome.not_care_home, 1704153600, PrimaryServiceType.non_residential, None, 50.0),
-        ("1-004", CareHome.not_care_home, 1704153600, PrimaryServiceType.non_residential, None, 60.0),
-        ("1-005", CareHome.care_home, 1704067200, PrimaryServiceType.care_home_only, 10, 4.0),
-        ("1-005", CareHome.not_care_home, 1704153600, PrimaryServiceType.non_residential, None, 50.0),
-    ]
-    expected_primary_service_rate_of_change_rows = [
-        ("1-001", CareHome.care_home, 1704067200, PrimaryServiceType.care_home_only, 10, 3.0, 1.0),
-        ("1-001", CareHome.care_home, 1704153600, PrimaryServiceType.care_home_only, 10, 2.8, 1.03999),
-        ("1-001", CareHome.care_home, 1704240000, PrimaryServiceType.care_home_only, 10, 3.4, 1.1176),
-        ("1-001", CareHome.care_home, 1704326400, PrimaryServiceType.care_home_only, 10, 3.2, 1.0854),
-        ("1-002", CareHome.care_home, 1704067200, PrimaryServiceType.care_home_only, 10, 2.0, 1.0),
-        ("1-002", CareHome.care_home, 1704153600, PrimaryServiceType.care_home_only, 10, None, 1.03999),
-        ("1-002", CareHome.care_home, 1704240000, PrimaryServiceType.care_home_only, 10, None, 1.1176),
-        ("1-002", CareHome.care_home, 1704326400, PrimaryServiceType.care_home_only, 10, 3.2, 1.0854),
-        ("1-003", CareHome.not_care_home, 1704067200, PrimaryServiceType.non_residential, None, 40.0, 1.0),
-        ("1-003", CareHome.not_care_home, 1704153600, PrimaryServiceType.non_residential, None, 50.0, 1.25),
-        ("1-004", CareHome.not_care_home, 1704153600, PrimaryServiceType.non_residential, None, 60.0, 1.25),
-        ("1-005", CareHome.care_home, 1704067200, PrimaryServiceType.care_home_only, 10, 4.0, 1.0),
-        ("1-005", CareHome.not_care_home, 1704153600, PrimaryServiceType.non_residential, None, 50.0, 1.25),
-    ]
-    # fmt: on
-
-    clean_column_with_values_rows = [
-        ("1-001", 1000000001, CareHome.care_home, 10.0),
-        ("1-001", 1000000002, CareHome.care_home, None),
-        ("1-001", 1000000003, CareHome.care_home, 10.0),
-    ]
-    expected_clean_column_with_values_rows = [
-        ("1-001", 1000000001, CareHome.care_home, 10.0, 1, 2),
-        ("1-001", 1000000002, CareHome.care_home, None, 1, 2),
-        ("1-001", 1000000003, CareHome.care_home, 10.0, 1, 2),
-    ]
-
-    clean_column_with_values_one_submission_rows = [
-        ("1-001", 1000000001, CareHome.care_home, 10.0),
-        ("1-001", 1000000002, CareHome.care_home, None),
-    ]
-    expected_clean_column_with_values_one_submission_rows = [
-        ("1-001", 1000000001, CareHome.care_home, None, 1, 1),
-        ("1-001", 1000000002, CareHome.care_home, None, 1, 1),
-    ]
-
-    clean_column_with_values_both_statuses_rows = [
-        ("1-001", 1000000001, CareHome.care_home, 10.0),
-        ("1-001", 1000000002, CareHome.care_home, 10.0),
-        ("1-001", 1000000003, CareHome.not_care_home, 10.0),
-    ]
-    expected_clean_column_with_values_both_statuses_rows = [
-        ("1-001", 1000000001, CareHome.care_home, None, 2, 2),
-        ("1-001", 1000000002, CareHome.care_home, None, 2, 2),
-        ("1-001", 1000000003, CareHome.not_care_home, None, 2, 1),
-    ]
-
-    calculate_care_home_status_count_rows = [
-        ("1-001", CareHome.care_home),
-        ("1-001", CareHome.care_home),
-        ("1-002", CareHome.care_home),
-        ("1-002", CareHome.not_care_home),
-    ]
-    expected_calculate_care_home_status_count_rows = [
-        ("1-001", CareHome.care_home, 1),
-        ("1-001", CareHome.care_home, 1),
-        ("1-002", CareHome.care_home, 2),
-        ("1-002", CareHome.not_care_home, 2),
-    ]
-
-    calculate_submission_count_same_care_home_status_rows = [
-        ("1-001", CareHome.care_home, None),
-        ("1-001", CareHome.care_home, None),
-        ("1-002", CareHome.care_home, None),
-        ("1-002", CareHome.care_home, 10.0),
-        ("1-003", CareHome.care_home, 10.0),
-        ("1-003", CareHome.care_home, 10.0),
-    ]
-    expected_calculate_submission_count_same_care_home_status_rows = [
-        ("1-001", CareHome.care_home, None, 0),
-        ("1-001", CareHome.care_home, None, 0),
-        ("1-002", CareHome.care_home, None, 1),
-        ("1-002", CareHome.care_home, 10.0, 1),
-        ("1-003", CareHome.care_home, 10.0, 2),
-        ("1-003", CareHome.care_home, 10.0, 2),
-    ]
-
-    calculate_submission_count_mixed_care_home_status_rows = [
-        ("1-001", CareHome.not_care_home, 10.0),
-        ("1-001", CareHome.care_home, 10.0),
-        ("1-001", CareHome.care_home, 10.0),
-    ]
-    expected_calculate_submission_count_mixed_care_home_status_rows = [
-        ("1-001", CareHome.not_care_home, 10.0, 1),
-        ("1-001", CareHome.care_home, 10.0, 2),
-        ("1-001", CareHome.care_home, 10.0, 2),
-    ]
-
-    interpolate_column_with_values_rows = [
-        ("1-001", 1704067200, 30.0),
-        ("1-001", 1704153600, None),
-        ("1-001", 1704240000, 34.0),
-        ("1-001", 1704326400, None),
-    ]
-    expected_interpolate_column_with_values_rows = [
-        ("1-001", 1704067200, 30.0, 30.0),
-        ("1-001", 1704153600, None, 32.0),
-        ("1-001", 1704240000, 34.0, 34.0),
-        ("1-001", 1704326400, None, None),
-    ]
-
-    add_previous_value_column_rows = [
-        ("1-001", 1672531200, 1.1),
-        ("1-001", 1672617600, 1.2),
-        ("1-001", 1672704000, 1.3),
-        ("1-001", 1672790400, 1.4),
-        ("1-002", 1672617600, 10.2),
-        ("1-002", 1672704000, 10.3),
-    ]
-    expected_add_previous_value_column_rows = [
-        ("1-001", 1672531200, 1.1, None),
-        ("1-001", 1672617600, 1.2, 1.1),
-        ("1-001", 1672704000, 1.3, 1.2),
-        ("1-001", 1672790400, 1.4, 1.3),
-        ("1-002", 1672617600, 10.2, None),
-        ("1-002", 1672704000, 10.3, 10.2),
-    ]
-
-    # fmt: off
-    add_rolling_sum_columns_rows = [
-        ("1-001", PrimaryServiceType.care_home_only, 1672531200, 1.1, None),
-        ("1-001", PrimaryServiceType.care_home_only, 1672617600, 1.2, 1.1),
-        ("1-001", PrimaryServiceType.care_home_only, 1672704000, 1.3, 1.2),
-        ("1-001", PrimaryServiceType.care_home_only, 1672790400, None, 1.3),
-        ("1-002", PrimaryServiceType.care_home_only, 1672531200, 1.4, None),
-        ("1-002", PrimaryServiceType.care_home_only, 1672617600, 1.3, 1.4),
-        ("1-003", PrimaryServiceType.non_residential, 1672531200, 10.0, None),
-        ("1-003", PrimaryServiceType.non_residential, 1672617600, 20.0, 10.0),
-        ("1-003", PrimaryServiceType.non_residential, 1672704000, 30.0, 20.0),
-    ]
-    expected_add_rolling_sum_columns_rows = [
-        ("1-001", PrimaryServiceType.care_home_only, 1672531200, 1.1, None, None, None),
-        ("1-001", PrimaryServiceType.care_home_only, 1672617600, 1.2, 1.1, 2.5, 2.5),
-        ("1-001", PrimaryServiceType.care_home_only, 1672704000, 1.3, 1.2, 3.8, 3.7),
-        ("1-001", PrimaryServiceType.care_home_only, 1672790400, None, 1.3, 3.8, 3.7),
-        ("1-002", PrimaryServiceType.care_home_only, 1672531200, 1.4, None, None, None),
-        ("1-002", PrimaryServiceType.care_home_only, 1672617600, 1.3, 1.4, 2.5, 2.5),
-        ("1-003", PrimaryServiceType.non_residential, 1672531200, 10.0, None, None, None),
-        ("1-003", PrimaryServiceType.non_residential, 1672617600, 20.0, 10.0, 20.0, 10.0),
-        ("1-003", PrimaryServiceType.non_residential, 1672704000, 30.0, 20.0, 50.0, 30.0),
-    ]
-    # fmt: on
-
-    calculate_rate_of_change_rows = [
-        ("1-001", 12.0, 10.0),
-        ("1-002", 15.0, None),
-        ("1-003", None, 20.0),
-        ("1-004", None, None),
-    ]
-    expected_calculate_rate_of_change_rows = [
-        ("1-001", 12.0, 10.0, 1.2),
-        ("1-002", 15.0, None, 1.0),
-        ("1-003", None, 20.0, 1.0),
-        ("1-004", None, None, 1.0),
-    ]
-
-
-@dataclass
-class ModelPrimaryServiceRateOfChangeTrendlineData:
-    # fmt: off
-    primary_service_rate_of_change_trendline_rows = [
-        ("1-001", 1704067200, CareHome.care_home, PrimaryServiceType.care_home_only, 3.0),
-        ("1-001", 1704153600, CareHome.care_home, PrimaryServiceType.care_home_only, 2.8),
-        ("1-001", 1704240000, CareHome.care_home, PrimaryServiceType.care_home_only, 3.4),
-        ("1-001", 1704326400, CareHome.care_home, PrimaryServiceType.care_home_only, 3.2),
-        ("1-002", 1704067200, CareHome.care_home, PrimaryServiceType.care_home_only, 2.0),
-        ("1-002", 1704153600, CareHome.care_home, PrimaryServiceType.care_home_only, None),
-        ("1-002", 1704240000, CareHome.care_home, PrimaryServiceType.care_home_only, None),
-        ("1-002", 1704326400, CareHome.care_home, PrimaryServiceType.care_home_only, 3.2),
-        ("1-003", 1704067200, CareHome.not_care_home, PrimaryServiceType.non_residential, 40.0),
-        ("1-003", 1704153600, CareHome.not_care_home, PrimaryServiceType.non_residential, 50.0),
-        ("1-004", 1704153600, CareHome.not_care_home, PrimaryServiceType.non_residential, 60.0),
-        ("1-005", 1704067200, CareHome.care_home, PrimaryServiceType.care_home_only, 4.0),
-        ("1-005", 1704153600, CareHome.not_care_home, PrimaryServiceType.non_residential, 50.0),
-    ]
-    expected_primary_service_rate_of_change_trendline_rows = [
-        ("1-001", 1704067200, CareHome.care_home, PrimaryServiceType.care_home_only, 3.0, 1.0),
-        ("1-001", 1704153600, CareHome.care_home, PrimaryServiceType.care_home_only, 2.8, 1.03999),
-        ("1-001", 1704240000, CareHome.care_home, PrimaryServiceType.care_home_only, 3.4, 1.16235),
-        ("1-001", 1704326400, CareHome.care_home, PrimaryServiceType.care_home_only, 3.2, 1.26158),
-        ("1-002", 1704067200, CareHome.care_home, PrimaryServiceType.care_home_only, 2.0, 1.0),
-        ("1-002", 1704153600, CareHome.care_home, PrimaryServiceType.care_home_only, None, 1.03999),
-        ("1-002", 1704240000, CareHome.care_home, PrimaryServiceType.care_home_only, None, 1.16235),
-        ("1-002", 1704326400, CareHome.care_home, PrimaryServiceType.care_home_only, 3.2, 1.26158),
-        ("1-003", 1704067200, CareHome.not_care_home, PrimaryServiceType.non_residential, 40.0, 1.0),
-        ("1-003", 1704153600, CareHome.not_care_home, PrimaryServiceType.non_residential, 50.0, 1.25),
-        ("1-004", 1704153600, CareHome.not_care_home, PrimaryServiceType.non_residential, 60.0, 1.25),
-        ("1-005", 1704067200, CareHome.care_home, PrimaryServiceType.care_home_only, 4.0, 1.0),
-        ("1-005", 1704153600, CareHome.not_care_home, PrimaryServiceType.non_residential, 50.0, 1.25),
-    ]
-    # fmt: on
-
-    calculate_rate_of_change_trendline_mock_rows = [
-        (PrimaryServiceType.care_home_only, 1672531200, 1.0),
-        (PrimaryServiceType.care_home_only, 1672617600, 1.5),
-        (PrimaryServiceType.care_home_only, 1672704000, 3.0),
-        (PrimaryServiceType.care_home_only, 1672790400, 4.5),
-        (PrimaryServiceType.non_residential, 1672531200, 1.0),
-        (PrimaryServiceType.non_residential, 1672617600, 1.2),
-        (PrimaryServiceType.non_residential, 1672704000, 1.2),
-        (PrimaryServiceType.non_residential, 1672790400, 1.8),
-    ]
-
-    deduplicate_dataframe_rows = [
-        (PrimaryServiceType.care_home_only, 1672531200, 1.0, 2.0),
-        (PrimaryServiceType.care_home_only, 1672617600, 1.1, 2.0),
-        (PrimaryServiceType.care_home_only, 1672704000, 1.2, 2.0),
-        (PrimaryServiceType.care_home_only, 1672790400, 1.3, 2.0),
-        (PrimaryServiceType.care_home_only, 1672531200, 1.0, 2.0),
-        (PrimaryServiceType.care_home_only, 1672617600, 1.1, 2.0),
-        (PrimaryServiceType.non_residential, 1672617600, 10.0, 2.0),
-        (PrimaryServiceType.non_residential, 1672617600, 10.0, 2.0),
-    ]
-    expected_deduplicate_dataframe_rows = [
-        (PrimaryServiceType.care_home_only, 1672531200, 1.0),
-        (PrimaryServiceType.care_home_only, 1672617600, 1.1),
-        (PrimaryServiceType.care_home_only, 1672704000, 1.2),
-        (PrimaryServiceType.care_home_only, 1672790400, 1.3),
-        (PrimaryServiceType.non_residential, 1672617600, 10.0),
-    ]
-
-    calculate_rate_of_change_trendline_rows = [
-        (PrimaryServiceType.care_home_only, 1672531200, 1.0),
-        (PrimaryServiceType.care_home_only, 1672617600, 1.5),
-        (PrimaryServiceType.care_home_only, 1672704000, 2.0),
-        (PrimaryServiceType.care_home_only, 1672790400, 1.5),
-        (PrimaryServiceType.non_residential, 1672531200, 1.0),
-        (PrimaryServiceType.non_residential, 1672617600, 1.2),
-        (PrimaryServiceType.non_residential, 1672704000, 1.0),
-        (PrimaryServiceType.non_residential, 1672790400, 1.5),
-    ]
-    expected_calculate_rate_of_change_trendline_rows = [
-        (PrimaryServiceType.care_home_only, 1672531200, 1.0),
-        (PrimaryServiceType.care_home_only, 1672617600, 1.5),
-        (PrimaryServiceType.care_home_only, 1672704000, 3.0),
-        (PrimaryServiceType.care_home_only, 1672790400, 4.5),
-        (PrimaryServiceType.non_residential, 1672531200, 1.0),
-        (PrimaryServiceType.non_residential, 1672617600, 1.2),
-        (PrimaryServiceType.non_residential, 1672704000, 1.2),
-        (PrimaryServiceType.non_residential, 1672790400, 1.8),
-    ]
-
-
-@dataclass
-class ModelRollingAverageData:
-    rolling_average_rows = [
-        ("1-001", 1672531200, 1.1),
-        ("1-001", 1672617600, 1.2),
-        ("1-001", 1672704000, 1.3),
-        ("1-001", 1672790400, 1.4),
-        ("1-001", 1672876800, 1.4),
-        ("1-001", 1672876800, 1.3),
-        ("1-002", 1672531200, 10.0),
-        ("1-002", 1672704000, 20.0),
-        ("1-002", 1672876800, 30.0),
-    ]
-    expected_rolling_average_rows = [
-        ("1-001", 1672531200, 1.1, 1.1),
-        ("1-001", 1672617600, 1.2, 1.15),
-        ("1-001", 1672704000, 1.3, 1.2),
-        ("1-001", 1672790400, 1.4, 1.3),
-        ("1-001", 1672876800, 1.4, 1.35),
-        ("1-001", 1672876800, 1.3, 1.35),
-        ("1-002", 1672531200, 10.0, 10.0),
-        ("1-002", 1672704000, 20.0, 15.0),
-        ("1-002", 1672876800, 30.0, 25.0),
-    ]
-
-
-@dataclass
-class ModelImputationWithExtrapolationAndInterpolationData:
-    imputation_with_extrapolation_and_interpolation_rows = [
-        ("1-001", date(2023, 1, 1), 1672531200, CareHome.care_home, 10.0, 15.0),
-        ("1-001", date(2023, 2, 1), 1675209600, CareHome.care_home, None, 15.1),
-        ("1-001", date(2023, 3, 1), 1677628800, CareHome.care_home, 30.0, 15.2),
-        ("1-002", date(2023, 1, 1), 1672531200, CareHome.not_care_home, None, 50.3),
-        ("1-002", date(2023, 2, 1), 1675209600, CareHome.not_care_home, 20.0, 50.5),
-        ("1-002", date(2023, 3, 1), 1677628800, CareHome.not_care_home, None, 50.7),
-        ("1-003", date(2023, 3, 1), 1677628800, CareHome.not_care_home, None, 50.7),
-    ]
-
-    split_dataset_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, True),
-        ("1-002", date(2024, 1, 1), CareHome.care_home, False),
-        ("1-003", date(2024, 1, 1), CareHome.not_care_home, True),
-        ("1-004", date(2024, 1, 1), CareHome.not_care_home, False),
-    ]
-    expected_split_dataset_imputation_df_when_true_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, True),
-    ]
-    expected_split_dataset_non_imputation_df_when_true_rows = [
-        ("1-002", date(2024, 1, 1), CareHome.care_home, False),
-        ("1-003", date(2024, 1, 1), CareHome.not_care_home, True),
-        ("1-004", date(2024, 1, 1), CareHome.not_care_home, False),
-    ]
-    expected_split_dataset_imputation_df_when_false_rows = [
-        ("1-003", date(2024, 1, 1), CareHome.not_care_home, True),
-    ]
-    expected_split_dataset_non_imputation_df_when_false_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, True),
-        ("1-002", date(2024, 1, 1), CareHome.care_home, False),
-        ("1-004", date(2024, 1, 1), CareHome.not_care_home, False),
-    ]
-
-    non_null_submission_when_locations_have_a_non_null_value_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, 10.0),
-        ("1-002", date(2024, 1, 1), CareHome.care_home, None),
-        ("1-002", date(2024, 2, 1), CareHome.care_home, 20.0),
-    ]
-    expected_non_null_submission_when_locations_have_a_non_null_value_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, 10.0, True),
-        ("1-002", date(2024, 1, 1), CareHome.care_home, None, True),
-        ("1-002", date(2024, 2, 1), CareHome.care_home, 20.0, True),
-    ]
-    non_null_submission_when_location_only_has_null_value_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, None),
-        ("1-001", date(2024, 2, 1), CareHome.care_home, None),
-    ]
-    expected_non_null_submission_when_location_only_has_null_value_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, None, False),
-        ("1-001", date(2024, 2, 1), CareHome.care_home, None, False),
-    ]
-    non_null_submission_when_a_location_has_both_care_home_options_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, None),
-        ("1-001", date(2024, 2, 1), CareHome.not_care_home, 30.0),
-    ]
-    expected_non_null_submission_when_a_location_has_both_care_home_options_rows = [
-        ("1-001", date(2024, 1, 1), CareHome.care_home, None, False),
-        ("1-001", date(2024, 2, 1), CareHome.not_care_home, 30.0, True),
-    ]
-
-    column_with_null_values_name: str = "null_values"
-    model_column_name: str = "trend_model"
-    imputation_model_column_name: str = "imputation_null_values_trend_model"
-
-    imputation_model_rows = [
-        ("1-001", None, None, None),
-        ("1-002", None, None, 30.0),
-        ("1-003", None, 30.0, None),
-        ("1-004", -2.0, None, None),
-        ("1-005", None, -2.0, 30.0),
-        ("1-006", 10.0, -2.0, None),
-        ("1-007", 10.0, None, 30.0),
-        ("1-008", 10.0, -2.0, 30.0),
-    ]
-    expected_imputation_model_rows = [
-        ("1-001", None, None, None, None),
-        ("1-002", None, None, 30.0, 30.0),
-        ("1-003", None, 30.0, None, 30.0),
-        ("1-004", -2.0, None, None, -2.0),
-        ("1-005", None, -2.0, 30.0, -2.0),
-        ("1-006", 10.0, -2.0, None, 10.0),
-        ("1-007", 10.0, None, 30.0, 10.0),
-        ("1-008", 10.0, -2.0, 30.0, 10.0),
-    ]
-
-
-@dataclass
-class ModelExtrapolation:
-    extrapolation_rows = [
-        ("1-001", date(2023, 1, 1), 1672531200, 15.0, 15.0),
-        ("1-001", date(2023, 2, 1), 1675209600, None, 15.1),
-        ("1-001", date(2023, 3, 1), 1677628800, 30.0, 15.2),
-        ("1-002", date(2023, 1, 1), 1672531200, 4.0, 50.3),
-        ("1-002", date(2023, 2, 1), 1675209600, None, 50.5),
-        ("1-002", date(2023, 3, 1), 1677628800, None, 50.7),
-        ("1-002", date(2023, 4, 1), 1680303600, None, 50.1),
-        ("1-003", date(2023, 1, 1), 1672531200, None, 50.3),
-        ("1-003", date(2023, 2, 1), 1675209600, 20.0, 50.5),
-        ("1-003", date(2023, 3, 1), 1677628800, None, 50.7),
-        ("1-004", date(2023, 3, 1), 1677628800, None, 50.7),
-    ]
-
-    first_and_last_submission_dates_rows = [
-        ("1-001", 1672531200, 15.0),
-        ("1-001", 1675209600, None),
-        ("1-001", 1677628800, 30.0),
-        ("1-002", 1672531200, None),
-        ("1-002", 1675209600, 4.0),
-        ("1-002", 1677628800, None),
-        ("1-003", 1677628800, None),
-    ]
-    expected_first_and_last_submission_dates_rows = [
-        ("1-001", 1672531200, 15.0, 1672531200, 1677628800),
-        ("1-001", 1675209600, None, 1672531200, 1677628800),
-        ("1-001", 1677628800, 30.0, 1672531200, 1677628800),
-        ("1-002", 1672531200, None, 1675209600, 1675209600),
-        ("1-002", 1675209600, 4.0, 1675209600, 1675209600),
-        ("1-002", 1677628800, None, 1675209600, 1675209600),
-        ("1-003", 1677628800, None, None, None),
-    ]
-
-    extrapolation_forwards_rows = [
-        ("1-001", 1672531200, 15.0, 10.0),
-        ("1-001", 1675209600, None, 20.0),
-        ("1-001", 1677628800, 30.0, 30.0),
-        ("1-002", 1672531200, None, 10.0),
-        ("1-002", 1675209600, 10.0, 20.0),
-        ("1-002", 1677628800, None, 30.0),
-        ("1-002", 1677629000, None, 100.0),
-        ("1-003", 1672531200, 20.0, 100.0),
-        ("1-003", 1675209600, None, 20.0),
-        ("1-004", 1677628800, None, 20.0),
-    ]
-    # fmt: off
-    expected_extrapolation_forwards_rows = [
-        ("1-001", 1672531200, 15.0, 10.0, None),
-        ("1-001", 1675209600, None, 20.0, 30.0),
-        ("1-001", 1677628800, 30.0, 30.0, 45.0),
-        ("1-002", 1672531200, None, 10.0, None),
-        ("1-002", 1675209600, 10.0, 20.0, None),
-        ("1-002", 1677628800, None, 30.0, 15.0),
-        ("1-002", 1677629000, None, 100.0, 40.0),  # capped at upper cutoff
-        ("1-003", 1672531200, 20.0, 100.0, None),
-        ("1-003", 1675209600, None, 20.0, 5.0),  # capped at lower cutoff
-        ("1-004", 1677628800, None, 20.0, None),
-    ]
-    # fmt: on
-    extrapolation_forwards_mock_rows = [
-        ("1-001", 12345, 15.0, 10.0, 15.0, 10.0),
-    ]
-
-    extrapolation_backwards_rows = [
-        ("1-001", 1672531200, 15.0, 1672531200, 1677628800, 10.0),
-        ("1-001", 1675209600, None, 1672531200, 1677628800, 20.0),
-        ("1-001", 1677628800, 30.0, 1672531200, 1677628800, 30.0),
-        ("1-002", 1672531200, None, 1675209600, 1675209600, 10.0),
-        ("1-002", 1675209600, 10.0, 1675209600, 1675209600, 20.0),
-        ("1-002", 1677628800, None, 1675209600, 1675209600, 30.0),
-        ("1-003", 1672531200, None, 1675209600, 1675209600, 1.0),
-        ("1-003", 1675209600, 20.0, 1675209600, 1675209600, 20.0),
-        ("1-004", 1672531200, None, 1675209600, 1675209600, 100.0),
-        ("1-004", 1675209600, 20.0, 1675209600, 1675209600, 20.0),
-        ("1-005", 1677628800, None, None, None, 20.0),
-    ]
-    # fmt: off
-    expected_extrapolation_backwards_rows = [
-        ("1-001", 1672531200, 15.0, 1672531200, 1677628800, 10.0, None),
-        ("1-001", 1675209600, None, 1672531200, 1677628800, 20.0, None),
-        ("1-001", 1677628800, 30.0, 1672531200, 1677628800, 30.0, None),
-        ("1-002", 1672531200, None, 1675209600, 1675209600, 10.0, 5.0),
-        ("1-002", 1675209600, 10.0, 1675209600, 1675209600, 20.0, None),
-        ("1-002", 1677628800, None, 1675209600, 1675209600, 30.0, None),
-        ("1-003", 1672531200, None, 1675209600, 1675209600, 1.0, 5.0),  # capped at lower cutoff
-        ("1-003", 1675209600, 20.0, 1675209600, 1675209600, 20.0, None),
-        ("1-004", 1672531200, None, 1675209600, 1675209600, 100.0, 80.0),  # capped at upper cutoff
-        ("1-004", 1675209600, 20.0, 1675209600, 1675209600, 20.0, None),
-        ("1-005", 1677628800, None, None, None, 20.0, None),
-    ]
-    # fmt: on
-    extrapolation_backwards_mock_rows = [
-        ("1-001", 12345, 15.0, 12345, 12345, 10.0, 15.0, 10.0),
-    ]
-
-    combine_extrapolation_rows = [
-        ("1-001", 1672531200, 15.0, 1672531200, 1677628800, 15.0, 15.0),
-        ("1-001", 1675209600, None, 1672531200, 1677628800, 20.0, 25.0),
-        ("1-001", 1677628800, 30.0, 1672531200, 1677628800, 30.0, 30.0),
-        ("1-002", 1672531200, None, 1675209600, 1675209600, 3.0, 2.0),
-        ("1-002", 1675209600, 4.0, 1675209600, 1675209600, 4.0, 4.0),
-        ("1-002", 1677628800, None, 1675209600, 1675209600, 5.0, 6.0),
-        ("1-003", 1677628800, None, None, None, None, None),
-    ]
-    expected_combine_extrapolation_rows = [
-        ("1-001", 1672531200, 15.0, 1672531200, 1677628800, 15.0, 15.0, None),
-        ("1-001", 1675209600, None, 1672531200, 1677628800, 20.0, 25.0, None),
-        ("1-001", 1677628800, 30.0, 1672531200, 1677628800, 30.0, 30.0, None),
-        ("1-002", 1672531200, None, 1675209600, 1675209600, 3.0, 2.0, 2.0),
-        ("1-002", 1675209600, 4.0, 1675209600, 1675209600, 4.0, 4.0, None),
-        ("1-002", 1677628800, None, 1675209600, 1675209600, 5.0, 6.0, 5.0),
-        ("1-003", 1677628800, None, None, None, None, None, None),
-    ]
-
-
-@dataclass
-class ModelInterpolation:
-    interpolation_rows = [
-        ("1-001", date(2023, 1, 1), 1672531200, None, None),
-        ("1-001", date(2023, 2, 1), 1675209600, None, None),
-        ("1-001", date(2023, 3, 1), 1677628800, 40.0, None),
-        ("1-001", date(2023, 4, 1), 1680307200, None, 42.0),
-        ("1-001", date(2023, 5, 1), 1682899200, None, 44.0),
-        ("1-001", date(2023, 6, 1), 1685577600, None, 46.0),
-        ("1-001", date(2023, 7, 1), 1688169600, None, 48.0),
-        ("1-001", date(2023, 8, 1), 1690848000, None, 50.0),
-        ("1-001", date(2023, 9, 1), 1693526400, None, 52.0),
-        ("1-001", date(2023, 10, 1), 1696118400, None, 54.0),
-        ("1-001", date(2023, 11, 1), 1698796800, None, 56.0),
-        ("1-001", date(2023, 12, 1), 1701388800, None, 58.0),
-        ("1-001", date(2024, 1, 1), 1704067200, None, 56.0),
-        ("1-001", date(2024, 2, 1), 1706745600, None, 54.0),
-        ("1-001", date(2024, 3, 1), 1709251200, 5.0, 52.0),
-        ("1-001", date(2024, 4, 1), 1711929600, None, 5.31),
-        ("1-001", date(2024, 5, 1), 1714521600, 15.0, 5.38),
-        ("1-001", date(2024, 6, 1), 1717200000, None, 13.93),
-    ]
-
-    calculate_residual_returns_none_when_extrapolation_forwards_is_none_rows = [
-        ("1-001", date(2023, 1, 1), 1672531200, None, None),
-        ("1-001", date(2023, 3, 1), 1677628800, 40.0, None),
-    ]
-    expected_calculate_residual_returns_none_when_extrapolation_forwards_is_none_rows = [
-        ("1-001", date(2023, 1, 1), 1672531200, None, None, None),
-        ("1-001", date(2023, 3, 1), 1677628800, 40.0, None, None),
-    ]
-    calculate_residual_returns_expected_values_when_extrapolation_forwards_is_known_rows = [
-        ("1-001", date(2023, 4, 1), 1680307200, None, 42.0),
-        ("1-001", date(2023, 5, 1), 1682899200, None, 44.0),
-        ("1-001", date(2024, 3, 1), 1709251200, 5.0, 52.0),
-        ("1-001", date(2024, 4, 1), 1711929600, None, 5.1),
-        ("1-001", date(2024, 5, 1), 1714521600, 15.0, 5.2),
-    ]
-    expected_calculate_residual_returns_expected_values_when_extrapolation_forwards_is_known_rows = [
-        ("1-001", date(2023, 4, 1), 1680307200, None, 42.0, -47.0),
-        ("1-001", date(2023, 5, 1), 1682899200, None, 44.0, -47.0),
-        ("1-001", date(2024, 3, 1), 1709251200, 5.0, 52.0, -47.0),
-        ("1-001", date(2024, 4, 1), 1711929600, None, 5.1, 9.8),
-        ("1-001", date(2024, 5, 1), 1714521600, 15.0, 5.2, 9.8),
-    ]
-    calculate_residual_returns_none_date_after_final_non_null_submission_rows = [
-        ("1-001", date(2024, 5, 1), 1714521600, 15.0, 5.2),
-        ("1-001", date(2024, 6, 1), 1717200000, None, 15.3),
-    ]
-    expected_calculate_residual_returns_none_date_after_final_non_null_submission_rows = [
-        ("1-001", date(2024, 5, 1), 1714521600, 15.0, 5.2, 9.8),
-        ("1-001", date(2024, 6, 1), 1717200000, None, 15.3, None),
-    ]
-
-    time_between_submissions_rows = [
-        ("1-001", date(2024, 2, 1), 1000000200, None),
-        ("1-001", date(2024, 3, 1), 1000000300, 5.0),
-        ("1-001", date(2024, 4, 1), 1000000400, None),
-        ("1-001", date(2024, 5, 1), 1000000500, None),
-        ("1-001", date(2024, 6, 1), 1000000600, None),
-        ("1-001", date(2024, 7, 1), 1000000700, 15.0),
-        ("1-001", date(2023, 8, 1), 1000000800, None),
-    ]
-    expected_time_between_submissions_rows = [
-        ("1-001", date(2024, 2, 1), 1000000200, None, None, None),
-        ("1-001", date(2024, 3, 1), 1000000300, 5.0, None, None),
-        ("1-001", date(2024, 4, 1), 1000000400, None, 400, 0.25),
-        ("1-001", date(2024, 5, 1), 1000000500, None, 400, 0.5),
-        ("1-001", date(2024, 6, 1), 1000000600, None, 400, 0.75),
-        ("1-001", date(2024, 7, 1), 1000000700, 15.0, None, None),
-        ("1-001", date(2023, 8, 1), 1000000800, None, None, None),
-    ]
-    time_between_submissions_mock_rows = [
-        ("1-001", date(2024, 2, 1), 12345, None, 12345, 12345),
-    ]
-
-    calculate_interpolated_values_rows = [
-        ("1-001", 172800, 20.0, None, None, None, None),
-        ("1-001", 259200, None, 20.0, 10.0, 345600, 0.25),
-        ("1-001", 345600, None, 20.0, 10.0, 345600, 0.5),
-        ("1-001", 432000, None, 20.0, 10.0, 345600, 0.75),
-        ("1-001", 518400, 30.0, 20.0, 10.0, None, None),
-        ("1-001", 604800, None, None, None, None, None),
-    ]
-    expected_calculate_interpolated_values_when_within_max_days_rows = [
-        ("1-001", 172800, 20.0, None, None, None, None, None),
-        ("1-001", 259200, None, 20.0, 10.0, 345600, 0.25, 22.5),
-        ("1-001", 345600, None, 20.0, 10.0, 345600, 0.5, 25.0),
-        ("1-001", 432000, None, 20.0, 10.0, 345600, 0.75, 27.5),
-        ("1-001", 518400, 30.0, 20.0, 10.0, None, None, None),
-        ("1-001", 604800, None, None, None, None, None, None),
-    ]
-    expected_calculate_interpolated_values_when_outside_of_max_days_rows = [
-        ("1-001", 172800, 20.0, None, None, None, None, None),
-        ("1-001", 259200, None, 20.0, 10.0, 345600, 0.25, None),
-        ("1-001", 345600, None, 20.0, 10.0, 345600, 0.5, None),
-        ("1-001", 432000, None, 20.0, 10.0, 345600, 0.75, None),
-        ("1-001", 518400, 30.0, 20.0, 10.0, None, None, None),
-        ("1-001", 604800, None, None, None, None, None, None),
-    ]
-
-
-@dataclass
-class ModelCareHomes:
-    care_homes_cleaned_ind_cqc_rows = [
-        (
-            "1-000000001",
-            "Care home with nursing",
-            None,
-            None,
-            "Y",
-            "South West",
-            67,
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000002",
-            "Care home without nursing",
-            None,
-            None,
-            "N",
-            "Merseyside",
-            12,
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000003",
-            "Care home with nursing",
-            None,
-            None,
-            None,
-            "Merseyside",
-            34,
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000004",
-            "non-residential",
-            10.0,
-            "already_populated",
-            "N",
-            None,
-            0,
-            date(2022, 3, 29),
-        ),
-        ("1-000000001", "non-residential", None, None, "N", None, 0, date(2022, 2, 20)),
-    ]
-    care_homes_features_rows = [
-        (
-            "1-000000001",
-            date(2022, 3, 29),
-            10,
-            62.0,
-            Vectors.sparse(39, {0: 1.0, 1: 1.2, 2: 1.0, 3: 50.0}),
-        ),
-        (
-            "1-000000003",
-            date(2022, 3, 29),
-            15,
-            45.0,
-            None,
-        ),
-    ]
-
-
-@dataclass
-class ModelNonResWithDormancy:
-    non_res_with_dormancy_cleaned_ind_cqc_rows = [
-        (
-            "1-000000001",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            "Y",
-            "South West",
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000002",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            "N",
-            "Merseyside",
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000003",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            None,
-            "Merseyside",
-            date(2022, 3, 29),
-        ),
-    ]
-    non_res_with_dormancy_features_rows = [
-        (
-            "1-000000001",
-            date(2022, 3, 29),
-            10.0,
-            Vectors.sparse(
-                32,
-                {
-                    0: 1.0,
-                    1: 1.0,
-                    4: 17.5,
-                    10: 1.0,
-                    18: 1.0,
-                    31: 35.0,
-                },
-            ),
-        ),
-        (
-            "1-000000003",
-            date(2022, 3, 29),
-            20.0,
-            None,
-        ),
-    ]
-
-
-@dataclass
-class ModelNonResWithoutDormancy:
-    non_res_without_dormancy_cleaned_ind_cqc_rows = [
-        (
-            "1-000000001",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            "Y",
-            "South West",
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000002",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            "N",
-            "Merseyside",
-            date(2022, 3, 29),
-        ),
-        (
-            "1-000000003",
-            PrimaryServiceType.non_residential,
-            None,
-            None,
-            None,
-            "Merseyside",
-            date(2022, 3, 29),
-        ),
-    ]
-    non_res_without_dormancy_features_rows = [
-        (
-            "1-000000001",
-            date(2022, 3, 29),
-            10.0,
-            Vectors.sparse(
-                31,
-                {
-                    0: 1.0,
-                    1: 1.0,
-                    3: 17.5,
-                    9: 1.0,
-                    17: 1.0,
-                    30: 35.0,
-                },
-            ),
-        ),
-        (
-            "1-000000003",
-            date(2022, 3, 29),
-            20.0,
-            None,
-        ),
-    ]
-
-
-@dataclass
-class ModelNonResWithAndWithoutDormancyCombinedRows:
-    estimated_posts_rows = [
-        ("1-001", date(2021, 1, 1), CareHome.not_care_home, "Y", 1, 1.0, None),
-        ("1-001", date(2022, 2, 1), CareHome.not_care_home, "Y", 2, 3.0, None),
-        ("1-001", date(2023, 3, 1), CareHome.not_care_home, "Y", 3, 4.0, 5.0),
-        ("1-001", date(2024, 4, 1), CareHome.not_care_home, "Y", 4, 5.0, 5.5),
-        ("1-001", date(2025, 5, 1), CareHome.not_care_home, "Y", 5, 6.0, 6.0),
-        ("1-001", date(2025, 6, 1), CareHome.not_care_home, "Y", 6, 7.0, 6.5),
-        ("1-002", date(2021, 1, 1), CareHome.not_care_home, "Y", 3, 8.0, None),
-        ("1-002", date(2022, 2, 1), CareHome.not_care_home, "Y", 4, 8.0, None),
-        ("1-002", date(2023, 3, 1), CareHome.not_care_home, "Y", 5, 8.0, 4.0),
-        ("1-002", date(2024, 4, 1), CareHome.not_care_home, "Y", 6, 8.0, 4.5),
-        ("1-002", date(2025, 5, 1), CareHome.not_care_home, "Y", 7, 8.0, 5.0),
-        ("1-002", date(2025, 6, 1), CareHome.not_care_home, "Y", 8, 8.0, 5.5),
-        ("1-003", date(2021, 1, 1), CareHome.not_care_home, "N", 1, 2.0, None),
-        ("1-003", date(2022, 2, 1), CareHome.not_care_home, "N", 2, 2.0, None),
-        ("1-003", date(2021, 3, 1), CareHome.not_care_home, "N", 3, 4.0, None),
-        ("1-003", date(2022, 4, 1), CareHome.not_care_home, "N", 4, 4.0, None),
-        ("1-003", date(2023, 5, 1), CareHome.not_care_home, "N", 5, 6.0, 8.0),
-        ("1-003", date(2024, 6, 1), CareHome.not_care_home, "N", 6, 6.0, 9.0),
-        ("1-004", date(2024, 4, 1), CareHome.care_home, "Y", 1, None, None),
-        ("1-005", date(2024, 5, 1), CareHome.not_care_home, "Y", 1, 4.0, 2.0),
-        ("1-005", date(2024, 6, 1), CareHome.not_care_home, "Y", 2, 5.0, 2.5),
-        ("1-006", date(2024, 5, 1), CareHome.not_care_home, "N", 1, 3.0, 2.5),
-        ("1-006", date(2024, 6, 1), CareHome.not_care_home, "N", 2, 3.0, 3.0),
-        ("1-006", date(2024, 7, 1), CareHome.not_care_home, "N", 3, 3.0, 3.0),
-        ("1-006", date(2024, 8, 1), CareHome.not_care_home, "N", 4, 3.0, 3.0),
-    ]
-
-    group_time_registered_to_six_month_bands_rows = [
-        ("1-001", 6),
-        ("1-002", 7),
-        ("1-003", 200),
-    ]
-    expected_group_time_registered_to_six_month_bands_rows = [
-        ("1-001", 6, 0),
-        ("1-002", 7, 1),
-        ("1-003", 200, 20),
-    ]
-
-    calculate_and_apply_model_ratios_rows = [
-        ("1-001", date(2022, 2, 1), "Y", 2, 3.0, None),
-        ("1-001", date(2023, 3, 1), "Y", 3, 4.0, 5.0),
-        ("1-002", date(2022, 2, 1), "Y", 4, 8.0, None),
-        ("1-002", date(2023, 3, 1), "Y", 5, 8.0, 4.0),
-        ("1-003", date(2022, 2, 1), "N", 2, 2.0, None),
-        ("1-003", date(2021, 3, 1), "N", 3, 4.0, None),
-        ("1-003", date(2022, 4, 1), "N", 4, 4.0, None),
-        ("1-003", date(2023, 5, 1), "N", 5, 6.0, 8.0),
-        ("1-003", date(2024, 6, 1), "N", 6, 6.0, 9.0),
-        ("1-004", date(2024, 5, 1), "Y", 1, 4.0, 2.0),
-        ("1-004", date(2024, 6, 1), "Y", 2, 5.0, 2.5),
-    ]
-
-    average_models_by_related_location_and_time_registered_rows = [
-        ("1-001", RelatedLocation.no_related_location, 1, 5.0, 14.0),
-        ("1-002", RelatedLocation.no_related_location, 1, 6.0, 15.0),
-        ("1-003", RelatedLocation.has_related_location, 1, 1.0, 10.0),
-        ("1-004", RelatedLocation.has_related_location, 1, 2.0, 11.0),
-        ("1-005", RelatedLocation.has_related_location, 2, 3.0, 12.0),
-        ("1-006", RelatedLocation.has_related_location, 2, 4.0, 13.0),
-        ("1-007", RelatedLocation.has_related_location, 2, 20.0, None),
-        ("1-008", RelatedLocation.has_related_location, 2, None, 20.0),
-    ]
-    expected_average_models_by_related_location_and_time_registered_rows = [
-        (RelatedLocation.no_related_location, 1, 5.5, 14.5),
-        (RelatedLocation.has_related_location, 1, 1.5, 10.5),
-        (RelatedLocation.has_related_location, 2, 3.5, 12.5),
-    ]
-
-    calculate_adjustment_ratios_rows = [
-        (RelatedLocation.no_related_location, 1, 5.0, 10.0),
-        (RelatedLocation.has_related_location, 1, 4.5, 1.5),
-    ]
-    expected_calculate_adjustment_ratios_rows = [
-        (RelatedLocation.no_related_location, 1, 5.0, 10.0, 0.5),
-        (RelatedLocation.has_related_location, 1, 4.5, 1.5, 3.0),
-    ]
-
-    calculate_adjustment_ratios_when_without_dormancy_is_zero_or_null_returns_one_rows = [
-        (RelatedLocation.no_related_location, 1, 5.0, 0.0),
-        (RelatedLocation.has_related_location, 1, 4.5, None),
-    ]
-    expected_calculate_adjustment_ratios_when_without_dormancy_is_zero_or_null_returns_one_rows = [
-        (RelatedLocation.no_related_location, 1, 5.0, 0.0, 1.0),
-        (RelatedLocation.has_related_location, 1, 4.5, None, 1.0),
-    ]
-
-    apply_model_ratios_returns_expected_values_when_all_values_known_rows = [
-        ("1-001", 5.0, 14.0, 0.25),
-        ("1-002", 6.0, 15.0, 2.0),
-    ]
-    expected_apply_model_ratios_returns_expected_values_when_all_values_known_rows = [
-        ("1-001", 5.0, 14.0, 0.25, 3.5),
-        ("1-002", 6.0, 15.0, 2.0, 30.0),
-    ]
-
-    apply_model_ratios_returns_none_when_none_values_present_rows = [
-        ("1-001", 5.0, None, 0.2),
-        ("1-002", 5.0, 10.0, None),
-        ("1-003", 5.0, None, None),
-    ]
-    expected_apply_model_ratios_returns_none_when_none_values_present_rows = [
-        ("1-001", 5.0, None, 0.2, None),
-        ("1-002", 5.0, 10.0, None, None),
-        ("1-003", 5.0, None, None, None),
-    ]
-
-    # fmt: off
-    calculate_and_apply_residuals_rows = [
-        ("1-001", date(2025, 2, 1), 20.0, 15.0),  # dates match, both models not null, residual calculated and applied
-        ("1-002", date(2025, 1, 1), None, 16.0),  # "1-002" - with_dormancy is null, residual added from date(2025, 2, 1) but not applied
-        ("1-002", date(2025, 2, 1), 10.0, 15.0),  # "1-002" - first period with both models present, take the residual
-        ("1-002", date(2025, 3, 1), 11.0, 14.0),  # "1-002" - residual added from date(2025, 2, 1)
-        ("1-002", date(2025, 4, 1), 12.0, None),  # "1-002" - without_dormancy is null, residual added from date(2025, 2, 1) but not applied
-        ("1-003", date(2025, 2, 1), 30.0, None),  # doesn't pass filter, no residual, keep original model value
-        ("1-004", date(2025, 2, 1), None, 15.0),  # doesn't pass filter, no residual, keep original model value
-        ("1-005", date(2025, 2, 1), None, None),  # doesn't pass filter, no residual, keep original model value
-    ]
-    expected_calculate_and_apply_residuals_rows = [
-        ("1-001", date(2025, 2, 1), 20.0, 15.0, 5.0, 20.0),  # dates match, both models not null, residual calculated
-        ("1-002", date(2025, 1, 1), None, 16.0, -5.0, 11.0),  # "1-002" - with_dormancy is null, residual added from date(2025, 2, 1) but not applied
-        ("1-002", date(2025, 2, 1), 10.0, 15.0, -5.0, 10.0),  # "1-002" - first period with both models present, take the residual
-        ("1-002", date(2025, 3, 1), 11.0, 14.0, -5.0, 9.0),  # "1-002" - residual added from date(2025, 2, 1)
-        ("1-002", date(2025, 4, 1), 12.0, None, -5.0, None),  # "1-002" - without_dormancy is null, residual added from date(2025, 2, 1) but not applied
-        ("1-003", date(2025, 2, 1), 30.0, None, None, None),  # doesn't pass filter, no residual, keep original model value
-        ("1-004", date(2025, 2, 1), None, 15.0, None, 15.0),  # doesn't pass filter, no residual, keep original model value
-        ("1-005", date(2025, 2, 1), None, None, None, None),  # doesn't pass filter, no residual, keep original model value
-    ]
-    # fmt: on
-
-    # fmt: off
-    calculate_residuals_rows = [
-        ("1-001", date(2025, 1, 1), date(2025, 2, 1), 10.0, 15.0),  # filtered out, dates not equal
-        ("1-002", date(2025, 2, 1), date(2025, 2, 1), 10.0, 15.0),  # not filtered, negative residual
-        ("1-003", date(2025, 2, 1), date(2025, 2, 1), 20.0, 15.0),  # not filtered, positive residual
-        ("1-004", date(2025, 2, 1), date(2025, 2, 1), 30.0, None),  # filtered out, null model value
-        ("1-005", date(2025, 2, 1), date(2025, 2, 1), None, 15.0),  # filtered out, null model value
-        ("1-006", date(2025, 2, 1), date(2025, 2, 1), None, None),  # filtered out, null model value
-    ]
-    expected_calculate_residuals_rows = [
-        ("1-002", -5.0),  # not filtered, negative residual
-        ("1-003", 5.0),  # not filtered, positive residual
-    ]
-    # fmt: on
-
-    apply_residuals_rows = [
-        ("1-001", 7.0, 12.0),
-        ("1-002", 5.0, -0.5),
-        ("1-003", 1.0, -2.5),
-        ("1-004", 10.0, None),
-        ("1-005", None, -1.0),
-        ("1-006", None, None),
-    ]
-    expected_apply_residuals_rows = [
-        ("1-001", 7.0, 12.0, 19.0),
-        ("1-002", 5.0, -0.5, 4.5),
-        ("1-003", 1.0, -2.5, -1.5),
-        ("1-004", 10.0, None, 10.0),
-        ("1-005", None, -1.0, None),
-        ("1-006", None, None, None),
-    ]
-
-    combine_model_predictions_rows = [
-        ("1-001", 10.0, 15.0),
-        ("1-002", 11.0, None),
-        ("1-003", None, 16.0),
-        ("1-004", None, None),
-    ]
-    expected_combine_model_predictions_rows = [
-        ("1-001", 10.0, 15.0, 10.0),
-        ("1-002", 11.0, None, 11.0),
-        ("1-003", None, 16.0, 16.0),
-        ("1-004", None, None, None),
-    ]
-
-
-@dataclass
-class EstimateFilledPostsModelsUtils:
-    cleaned_cqc_rows = ModelCareHomes.care_homes_cleaned_ind_cqc_rows
-
-    predictions_rows = [
-        (
-            "1-000000001",
-            "Care home with nursing",
-            50.0,
-            "Y",
-            "South West",
-            67,
-            date(2022, 3, 29),
-            56.89,
-        ),
-    ]
-
-    set_min_value_when_below_minimum_rows = [
-        ("1-001", 0.5, -7.5),
-    ]
-    expected_set_min_value_when_below_min_value_rows = [
-        ("1-001", 0.5, 2.0),
-    ]
-    expected_set_min_value_when_below_minimum_and_default_not_set_rows = [
-        ("1-001", 0.5, 1.0),
-    ]
-    expected_set_min_value_when_below_minimum_and_min_value_is_negative_rows = [
-        ("1-001", 0.5, -5.0),
-    ]
-
-    set_min_value_when_above_minimum_rows = [
-        ("1-001", 1.5, 1.5),
-    ]
-
-    set_min_value_when_null_rows = [
-        ("1-001", None, None),
-    ]
-
-    combine_care_home_ratios_and_non_res_posts_rows = [
-        ("1-001", CareHome.care_home, 20.0, 1.6),
-        ("1-002", CareHome.care_home, 10.0, None),
-        ("1-003", CareHome.care_home, None, 1.8),
-        ("1-004", CareHome.care_home, None, None),
-        ("1-005", CareHome.not_care_home, 20.0, 1.6),
-        ("1-006", CareHome.not_care_home, 10.0, None),
-        ("1-007", CareHome.not_care_home, None, 1.6),
-        ("1-008", CareHome.not_care_home, None, None),
-    ]
-    expected_combine_care_home_ratios_and_non_res_posts_rows = [
-        ("1-001", CareHome.care_home, 20.0, 1.6, 1.6),
-        ("1-002", CareHome.care_home, 10.0, None, None),
-        ("1-003", CareHome.care_home, None, 1.8, 1.8),
-        ("1-004", CareHome.care_home, None, None, None),
-        ("1-005", CareHome.not_care_home, 20.0, 1.6, 20.0),
-        ("1-006", CareHome.not_care_home, 10.0, None, 10.0),
-        ("1-007", CareHome.not_care_home, None, 1.8, None),
-        ("1-008", CareHome.not_care_home, None, None, None),
-    ]
-
-    clean_number_of_beds_banded_rows = [
-        ("1-001", PrimaryServiceType.care_home_only, 1.0),
-        ("1-002", PrimaryServiceType.care_home_only, 2.0),
-        ("1-003", PrimaryServiceType.care_home_only, 3.0),
-        ("1-004", PrimaryServiceType.care_home_only, 4.0),
-        ("1-005", PrimaryServiceType.care_home_only, None),
-        ("1-006", PrimaryServiceType.care_home_with_nursing, 1.0),
-        ("1-007", PrimaryServiceType.care_home_with_nursing, 2.0),
-        ("1-008", PrimaryServiceType.care_home_with_nursing, 3.0),
-        ("1-009", PrimaryServiceType.care_home_with_nursing, 4.0),
-        ("1-010", PrimaryServiceType.care_home_with_nursing, None),
-        ("1-011", PrimaryServiceType.non_residential, 1.0),
-        ("1-012", PrimaryServiceType.non_residential, None),
-    ]
-    expected_clean_number_of_beds_banded_rows = [
-        ("1-001", PrimaryServiceType.care_home_only, 1.0, 2.0),
-        ("1-002", PrimaryServiceType.care_home_only, 2.0, 2.0),
-        ("1-003", PrimaryServiceType.care_home_only, 3.0, 3.0),
-        ("1-004", PrimaryServiceType.care_home_only, 4.0, 4.0),
-        ("1-005", PrimaryServiceType.care_home_only, None, None),
-        ("1-006", PrimaryServiceType.care_home_with_nursing, 1.0, 3.0),
-        ("1-007", PrimaryServiceType.care_home_with_nursing, 2.0, 3.0),
-        ("1-008", PrimaryServiceType.care_home_with_nursing, 3.0, 3.0),
-        ("1-009", PrimaryServiceType.care_home_with_nursing, 4.0, 4.0),
-        ("1-010", PrimaryServiceType.care_home_with_nursing, None, None),
-        ("1-011", PrimaryServiceType.non_residential, 1.0, 1.0),
-        ("1-012", PrimaryServiceType.non_residential, None, None),
-    ]
-
-    convert_care_home_ratios_to_filled_posts_and_merge_with_filled_post_values_rows = [
-        ("1-001", CareHome.care_home, 5, 1.6, 20.0),
-        ("1-002", CareHome.care_home, 5, None, 10.0),
-        ("1-003", CareHome.care_home, None, 1.6, 20.0),
-        ("1-004", CareHome.care_home, None, None, 10.0),
-        ("1-005", CareHome.care_home, 5, 1.8, None),
-        ("1-006", CareHome.care_home, 5, None, None),
-        ("1-007", CareHome.care_home, None, 1.8, None),
-        ("1-008", CareHome.care_home, None, None, None),
-        ("1-009", CareHome.not_care_home, 5, 1.6, 20.0),
-        ("1-010", CareHome.not_care_home, None, None, 10.0),
-        ("1-011", CareHome.not_care_home, 5, 1.6, None),
-        ("1-012", CareHome.not_care_home, None, None, None),
-    ]
-    expected_convert_care_home_ratios_to_filled_posts_and_merge_with_filled_post_values_rows = [
-        ("1-001", CareHome.care_home, 5, 1.6, 8.0),
-        ("1-002", CareHome.care_home, 5, None, None),
-        ("1-003", CareHome.care_home, None, 1.6, None),
-        ("1-004", CareHome.care_home, None, None, None),
-        ("1-005", CareHome.care_home, 5, 1.8, 9.0),
-        ("1-006", CareHome.care_home, 5, None, None),
-        ("1-007", CareHome.care_home, None, 1.8, None),
-        ("1-008", CareHome.care_home, None, None, None),
-        ("1-009", CareHome.not_care_home, 5, 1.6, 20.0),
-        ("1-010", CareHome.not_care_home, None, None, 10.0),
-        ("1-011", CareHome.not_care_home, 5, 1.6, None),
-        ("1-012", CareHome.not_care_home, None, None, None),
-    ]
-
-    create_test_and_train_datasets_rows = [
-        ("1-001", Vectors.dense([10.0, 0.0, 1.0])),
-        ("1-002", Vectors.dense([20.0, 1.0, 1.0])),
-        ("1-003", Vectors.dense([30.0, 0.0, 1.0])),
-        ("1-004", Vectors.dense([40.0, 0.0, 1.0])),
-        ("1-005", Vectors.dense([50.0, 1.0, 1.0])),
-    ]
-
-    train_lasso_regression_model_rows = [
-        (Vectors.dense([1.0, 2.0]), 5.0),
-        (Vectors.dense([2.0, 1.0]), 4.0),
-    ]
-
-
-@dataclass
-class MLModelMetrics:
-    ind_cqc_with_predictions_rows = [
-        ("1-00001", "care home", 50.0, "Y", "South West", 67, date(2022, 3, 9), 56.89),
-        ("1-00002", "non-res", 10.0, "N", "North East", 0, date(2022, 3, 9), 12.34),
-    ]
-
-    r2_metric_rows = [
-        ("1-00001", 50.0, 56.89),
-        ("1-00002", 10.0, 12.34),
-    ]
-
-    predictions_rows = [
-        ("1-00001", 50.0, 56.89),
-        ("1-00002", None, 46.80),
-        ("1-00003", 10.0, 12.34),
-    ]
-    expected_predictions_with_dependent_rows = [
-        ("1-00001", 50.0, 56.89),
-        ("1-00003", 10.0, 12.34),
-    ]
-
-
-@dataclass
 class ValidationUtils:
     size_of_dataset_rule = {RuleName.size_of_dataset: 3}
     size_of_dataset_success_rows = [
